flask/requirements.txt
Flask
python-dotenv
psycopg2-binary
SQLAlchemy
Flask-SQLAlchemy
Werkzeug
flask-wtf
playwright
